+++ src/test/java/cambio/simulator/entities/patterns/CircuitBreakerTest.java
@@ -5,9 +5,9 @@ import java.util.Set;
 
 import cambio.simulator.entities.microservice.Microservice;
 import cambio.simulator.entities.microservice.MicroserviceInstance;
-import cambio.simulator.testutils.RandomTieredModel;
-import cambio.simulator.testutils.TestExperiment;
-import cambio.simulator.testutils.TestUtils;
+import cambio.simulator.test.RandomTieredModel;
+import cambio.simulator.test.TestExperiment;
+import cambio.simulator.test.TestUtils;
 import desmoj.core.simulator.Experiment;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
