+++ src/test/java/cambio/simulator/resources/CPUProcessSchedulerTest.java
@@ -17,8 +17,8 @@ import java.util.Random;
 
 import cambio.simulator.resources.cpu.CPUProcess;
 import cambio.simulator.resources.cpu.scheduling.CPUProcessScheduler;
-import cambio.simulator.testutils.TestExperiment;
-import cambio.simulator.testutils.TestModel;
+import cambio.simulator.test.TestExperiment;
+import cambio.simulator.test.TestModel;
 import org.javatuples.Pair;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Assumptions;
