+++ src/test/java/cambio/simulator/scenarios/ChaosMonkeyExperimentTest.java
@@ -3,7 +3,7 @@ package cambio.simulator.scenarios;
 import java.io.File;
 
 import cambio.simulator.Main;
-import cambio.simulator.testutils.FileLoaderUtil;
+import cambio.simulator.test.FileLoaderUtil;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
