+++ src/test/java/cambio/simulator/test/TestScheduler.java
@@ -1,4 +1,4 @@
-package cambio.simulator.testutils;
+package cambio.simulator.test;
 
 import desmoj.core.simulator.EventList;
 import desmoj.core.simulator.Experiment;
